physical-ai-textbook/chatbot/requirements.txt
# Core Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Vector Database
qdrant-client==1.12.1

# Embeddings & LLM
cohere==5.14.0
openai==1.57.4

# Environment Management
python-dotenv==1.0.1

# Data Processing
langchain==0.3.13
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-cohere==0.5.0

# Text Processing & Chunking
tiktoken==0.8.0
markdown==3.7

# HTTP & Async
httpx==0.28.1
aiofiles==24.1.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.0
typing-extensions==4.12.2

# Optional: Progress & Logging
tqdm==4.67.1
loguru==0.7.3

# Optional: RAG Frameworks (if needed)
llama-index==0.12.9
